[Refactoring] Refactor Mod command
